Refactor: 중복되는 reducer 파일 삭제하고, 하나의 reducer를 사용
사이드바에서 마일스톤, 어사이니, 라벨의 리듀서 로직이 모두 같기 때문에,
하나의 리듀서를 세 곳에서 공통으로 사용하도록 하였다.
